in progress: fixing - changing balance variable from double to BigDecimal. todo: fix tests after var type change

+++ src/Models/ClientAccount.java
@@ -1,21 +1,23 @@
 package Models;
 
+import java.math.BigDecimal;
+
 public class ClientAccount {
     private long id;
     private String name;
-    private double balance;
+    private BigDecimal balance;
 
     public String getName() {
         return name;
     }
 
-    public double getBalance() {
+    public BigDecimal getBalance() {
         return balance;
     }
 
     public long getId() { return id; }
 
-    public ClientAccount(long id, String name, double balance) {
+    public ClientAccount(long id, String name, BigDecimal balance) {
         this.id = id;
         this.name = name;
         this.balance = balance;
